add Parts Replacement report and maintenance pages; update routing and layout
